ShelfView: Load books by clicking on them
and removed the items shenaningans that was happening in ShelfView. at first
that really confused me. seems i made that in order to test the layout;
temporarily made it instantiate the same book 10 times in order to see how it
would look when you have a lot of books

and in fact, as i’m writing this i think i discovered a bug in the layout when
you only have a few items (there seem to be ghost items?)