BookView: Add skip line toolbar action
I also renamed setLine to _setLine because it’s quite an internal function; if
you call it on its own you’ll break everything as all the positions won’t be
set properly.

And I also changed the advanceLine logic in order to allow for calling it while
in the middle of lines, and still have it skip to the start of the next line.