HighlightingService: Fix returning if highlighting variables are ..
.. unset. Since setting persistent_pos to None on init that broke that. But,
still keeping getattr just in case it is not set at all.

the reason i was looking at this is to see what the default spacer sizing looks
like if modeline’s update_ is never called; i.e. when you switch to book view
with no book loaded. So I also attach in this commit my resulting change of the
spacer’s initial size.