BookView: Change centreAroundCursor approach to not rely on blocks
Qt seems to break the html input into QTextBrowser into blocks according to <p>
tags. Some books use <br/> instead of <p> -- causing Qt to create only one
large block. The previous approach centred around the block the cursor is in,
which causes the cursor to be out of view in such cases.