Fix bug to do with saving and restoring BookView’s splitter state
If closing retype without ever loading a book, since StatsDock would not get
added to the splitter in that situation, it would save the splitter with the
StatsDock in a collapsed state.

Fixed by adding StatsDock to the splitter from the start, and only connecting
it to the console after a book is loaded.