Highlighting: Separate highlight cursor
Tables and moving cursors with KeepAnchor do not get along. Cursor goes way off
the table. Having the cursor that does the highlighting be separate means main
cursor does not use KeepAnchor and can stay in the right position, though
tables still cause issues as the highlighting cursor then goes all the way to
the end, highlighting the whole table. I could not find a way to work around
that (other than hacky movement things that sometimes break) so this will
remain a known issue for now.