fix: BookDisplay hotkeys
When keys are typed in the BookDisplay, the focus and keyPressEvent is
transferred to the console under the assumption this was done accidentally and
the intent is to type along to the book. However, this also prevented
BookDisplay hotkeys like Ctrl+C to copy text.

So now only "naked" key presses, with no modifier keys held, will transfer to
the console.