feat(ui): Use custom widget LineEdit instead of QLineEdit
QLineEdit has some limitations; in particular, it is difficult to modify the
undo behaviour. The new widget LineEdit derives from QPlainTextEdit and has
better undo behaviour and also it will be easier to modify the undo stack
for it in future should that be necessary. QLineEdit does not allow you to do
that and the only ways to modify the undo actions is very hacky (like event
filtering the undo key combination, but then it does not apply to undo via
context menu, among other issues)