refactor(Library): Save entire save
Previously save was loading the contents of the save file before
modifying and saving a particular entry. This changes it to save the
entire save have instead. This will prevent duplicate entries after
v1 migration.

Also tests for the past 3 commits. They were written concurrently
but because I made the separate commits retroactively instead of one
big commit, adding the correct changes to the tests to each commit
would have been too much of a pain so updating them in one commit.