fix(Logging): Silence third-party warnings when not on DEBUG level
Currently ebooklib's dependency lxml emits a FutureWarning due to a
line that has to be fixed in ebooklib. I have made a PR, but until it
is merged and included in a release, we have this warning polluting
stdout.

Warnings emitted by dependencies are not a concern for normal users
anyway, so I thought it is more proper to filter them out when running
retype with a loglevel other than DEBUG.