fix: Handle duplicate themes/iconsets/wordlists in a clearer way
For themes/iconsets/wordlists that exist with the same name both in
root and in user dir, previously duplicates were ignored and only
the ones loaded first shown.

This causes confusion, for example in the situation where the user
dir is set to the application dir path of a previous version of
retype.

Instead, load the ones in application dir first, then if user dir
contains duplicates of ones already loaded, they are loaded anyway
with '(user dir)' suffix appended to their names.