fix: Handle the case of ebooklib failing to read the epub
Fixes #41... technically. It should not crash anymore, but it skips
the problematic book. Since it's ebooklib that handles the epub
reading, if an error gets raised that it doesn't catch, the book
will not be loaded properly, and there's nothing we can really do
about it on our end -- short of hacking the epub beforehand I guess
to avoid the error being raised in the first place, but at this
point might as well create our own epub library... which is beyond
the scope of retype.

With these changes there isn't an error dialog, just an error in the
log, with the message and also full stacktrace in case it's helpful
to diagnose what's wrong with the epub. We could guess what's wrong
in order to give a more intuitive indication to the user, but I'd
rather not go into that, I don't know all cases. And there could
also be cases where it's simply a bug in ebooklib or one of its
dependencies.

One potential thing that could be done is ask ebooklib to add an
option to be more permissive in epub reading and ignore it if the
manifest declares files that are not really there for example.

Also note that although the book is excluded from display in shelf
view, this is after a LibraryItem and idn were generated, so the
result is a jump in the idns, for instance if it was 1 the books
will be 0 2 3 4 (missing 1).