fix: Autosave erroneous error log
When closing retype without opening a book, it would log an error
that maybeSave was called without positioning variables set, which
is not actually an error in this case and is what is expected to
happen. I moved the positioning variables checks to inside the
check for self.book so that if a book was not opened in the session
nothing happens.